src/pyproject.toml
[tool.poetry]
name = "heygptcli"
version = "0.1.8"
description = ""
authors = ["Rishang <rishang@localhost.com>"]
readme = "README.md"
packages = [
  { include = "heygpt" }
]

[tool.poetry.scripts]
heygpt = "heygpt.cli:app"

[tool.poetry.dependencies]
python = "^3.10"
requests = "*"
fastapi = "*"
uvicorn = "*"
typer = "*"
python-dotenv = "^1.0.0"
openai = "^0.27.8"
rich = "^13.4.1"
prompt-toolkit = "^3.0.38"
streamlit = "^1.25.0"
google-generativeai = "^0.1.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.2"
black = "^23.3.0"
mypy = "^1.3.0"
types-requests = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
